spark_consumer/requirements.txt
pandas
numpy
scikit-learn
joblib
pyarrow
kafka-python==2.0.2
pyspark==3.4.0
psycopg2-binary
